pyproject.toml
[project]
name = "rpg-ia-master"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "chromadb>=1.3.5",
    "faiss-cpu>=1.13.1",
    "google-genai>=1.53.0",
    "graphviz>=0.21",
    "kuzu>=0.11.3",
    "langchain>=1.1.2",
    "langchain-chroma>=1.0.0",
    "langchain-community>=0.4.1",
    "langchain-experimental>=0.4.0",
    "langchain-google-genai>=3.2.0",
    "langgraph>=1.0.4",
    "langsmith>=0.4.55",
    "pydantic>=2.12.5",
    "python-dotenv>=1.2.1",
    "streamlit>=1.52.1",
]
